willway_git/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
python-dotenv==1.0.0
sqlalchemy==2.0.23
Werkzeug==2.3.7
requests==2.31.0
python-telegram-bot==20.7
gunicorn==21.2.0
